requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14.0
drf-spectacular>=0.27.0
psycopg2-binary>=2.9.0
djangorestframework-simplejwt>=5.3.0

